refactor: extract SynthContext types, add ErrorBoundary, improve arpeggiator timing
- Extract SynthContextType and useSynth hook from SynthContext for cleaner separation
- Add AudioErrorBoundary component to gracefully handle audio initialization errors
- Add dedicated playArpNote/stopArpNote functions with proper time scheduling
- Fix arpeggiator to use time-based Tone.js scheduling instead of Draw callbacks
- Improve LFO filter modulation range for more pronounced effect
- Initialize lowpass filter with frequency: 0 so Signal is sole source
- Clean up module imports to use barrel exports from context
- Add cleanup handler for arpeggiator unmount
