+++ src/App.tsx
@@ -1,12 +1,15 @@
 import { Synth } from './components/Synth'
 import { SynthProvider } from './context'
+import { AudioErrorBoundary } from './components/ErrorBoundary'
 import { getFrequency } from './constants/frequencies'
 
 function App() {
   return (
-    <SynthProvider getFrequency={getFrequency}>
-      <Synth />
-    </SynthProvider>
+    <AudioErrorBoundary>
+      <SynthProvider getFrequency={getFrequency}>
+        <Synth />
+      </SynthProvider>
+    </AudioErrorBoundary>
   )
 }
