+++ src/components/Modules/DistortionModule.tsx
@@ -1,3 +1,4 @@
+import { memo } from 'react'
 import { ModulePanel } from '../Layout/ModulePanel'
 import { Knob } from '../Controls'
 import { DEFAULT_EFFECT_PARAMS } from '../../types/synth.types'
@@ -10,7 +11,7 @@ interface DistortionModuleProps {
   className?: string
 }
 
-export function DistortionModule({
+export const DistortionModule = memo(function DistortionModule({
   amount,
   wet,
   onAmountChange,
@@ -39,4 +40,4 @@ export function DistortionModule({
       />
     </ModulePanel>
   )
-}
+})
