+++ src/components/Modules/ReverbModule.tsx
@@ -1,3 +1,4 @@
+import { memo } from 'react'
 import { ModulePanel } from '../Layout/ModulePanel'
 import { Knob } from '../Controls'
 import { DEFAULT_EFFECT_PARAMS } from '../../types/synth.types'
@@ -10,7 +11,7 @@ interface ReverbModuleProps {
   className?: string
 }
 
-export function ReverbModule({
+export const ReverbModule = memo(function ReverbModule({
   decay,
   wet,
   onDecayChange,
@@ -40,4 +41,4 @@ export function ReverbModule({
       />
     </ModulePanel>
   )
-}
+})
