+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/controls/knob.tsx","./src/components/controls/modwheel.tsx","./src/components/controls/octavedisplay.tsx","./src/components/controls/pitchwheel.tsx","./src/components/controls/segmentedselector.tsx","./src/components/controls/slider.tsx","./src/components/controls/togglebutton.tsx","./src/components/controls/waveformselector.tsx","./src/components/controls/index.ts","./src/components/keyboard/key.tsx","./src/components/keyboard/keyboard.tsx","./src/components/keyboard/index.ts","./src/components/layout/modulepanel.tsx","./src/components/layout/index.ts","./src/components/modules/arpeggiatormodule.tsx","./src/components/modules/chorusmodule.tsx","./src/components/modules/delaymodule.tsx","./src/components/modules/distortionmodule.tsx","./src/components/modules/filtermodule.tsx","./src/components/modules/lfomodule.tsx","./src/components/modules/mastermodule.tsx","./src/components/modules/oscillatormodule.tsx","./src/components/modules/phasermodule.tsx","./src/components/modules/pitchmodule.tsx","./src/components/modules/reverbmodule.tsx","./src/components/modules/tempomodule.tsx","./src/components/modules/index.ts","./src/components/presets/presetmanager.tsx","./src/components/presets/presetmodal.tsx","./src/components/presets/index.ts","./src/components/synth/synth.tsx","./src/components/synth/index.ts","./src/components/visualizers/spectrumanalyzer.tsx","./src/components/visualizers/vumeter.tsx","./src/components/visualizers/waveformdisplay.tsx","./src/components/visualizers/index.ts","./src/constants/factorypresets.ts","./src/constants/frequencies.ts","./src/context/synthcontext.tsx","./src/context/index.ts","./src/hooks/useanimationframe.ts","./src/hooks/usearpeggiator.ts","./src/hooks/useaudioengine.ts","./src/hooks/usekeyboard.ts","./src/hooks/usepresets.ts","./src/hooks/usetempo.ts","./src/types/synth.types.ts"],"version":"5.6.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/controls/knob.tsx","./src/components/controls/modwheel.tsx","./src/components/controls/octavedisplay.tsx","./src/components/controls/pitchwheel.tsx","./src/components/controls/segmentedselector.tsx","./src/components/controls/slider.tsx","./src/components/controls/togglebutton.tsx","./src/components/controls/waveformselector.tsx","./src/components/controls/index.ts","./src/components/errorboundary/audioerrorboundary.tsx","./src/components/errorboundary/index.ts","./src/components/keyboard/key.tsx","./src/components/keyboard/keyboard.tsx","./src/components/keyboard/index.ts","./src/components/layout/modulepanel.tsx","./src/components/layout/index.ts","./src/components/modules/arpeggiatormodule.tsx","./src/components/modules/chorusmodule.tsx","./src/components/modules/delaymodule.tsx","./src/components/modules/distortionmodule.tsx","./src/components/modules/filtermodule.tsx","./src/components/modules/lfomodule.tsx","./src/components/modules/mastermodule.tsx","./src/components/modules/oscillatormodule.tsx","./src/components/modules/phasermodule.tsx","./src/components/modules/pitchmodule.tsx","./src/components/modules/reverbmodule.tsx","./src/components/modules/tempomodule.tsx","./src/components/modules/index.ts","./src/components/presets/presetmanager.tsx","./src/components/presets/presetmodal.tsx","./src/components/presets/index.ts","./src/components/synth/synth.tsx","./src/components/synth/index.ts","./src/components/visualizers/spectrumanalyzer.tsx","./src/components/visualizers/vumeter.tsx","./src/components/visualizers/waveformdisplay.tsx","./src/components/visualizers/index.ts","./src/constants/factorypresets.ts","./src/constants/frequencies.ts","./src/context/synthcontext.tsx","./src/context/synthcontexttype.ts","./src/context/index.ts","./src/context/usesynth.ts","./src/hooks/useanimationframe.ts","./src/hooks/usearpeggiator.ts","./src/hooks/useaudioengine.ts","./src/hooks/usekeyboard.ts","./src/hooks/usepresets.ts","./src/hooks/usetempo.ts","./src/types/synth.types.ts"],"version":"5.6.3"}
